+++ CHANGELOG.md
@@ -4,6 +4,63 @@ All notable changes to Connector Service will be documented here.
 
 - - -
 
+## 2026.01.13.2
+
+### Bug Fixes
+
+- Adyen shoppername to none for bankredirect, repeatpayment ([#423](https://github.com/juspay/connector-service/pull/423)) ([`938d6f2`](https://github.com/juspay/connector-service/commit/938d6f2b9d0e5f76a5dec2320e4b35f455c61633))
+
+**Full Changelog:** [`2026.01.13.1...2026.01.13.2`](https://github.com/juspay/connector-service/compare/2026.01.13.1...2026.01.13.2)
+
+- - -
+
+## 2026.01.13.1
+
+### Bug Fixes
+
+- RepeatPayment Merchant configured Currency Handling ([#422](https://github.com/juspay/connector-service/pull/422)) ([`4cef8ef`](https://github.com/juspay/connector-service/commit/4cef8ef84ede9adac78eac27c557d161af36f306))
+
+**Full Changelog:** [`2026.01.13.0...2026.01.13.1`](https://github.com/juspay/connector-service/compare/2026.01.13.0...2026.01.13.1)
+
+- - -
+
+## 2026.01.13.0
+
+### Features
+
+- **connector:** [GETNETGLOBAL] Connector Integration ([#381](https://github.com/juspay/connector-service/pull/381)) ([`a648f45`](https://github.com/juspay/connector-service/commit/a648f45d93309a083c90f59f73d4f2f5a95effb1))
+
+**Full Changelog:** [`2026.01.12.1...2026.01.13.0`](https://github.com/juspay/connector-service/compare/2026.01.12.1...2026.01.13.0)
+
+- - -
+
+## 2026.01.12.1
+
+### Bug Fixes
+
+- **connector:** Fix Razorpay metadata to accept all values ([#418](https://github.com/juspay/connector-service/pull/418)) ([`4006a76`](https://github.com/juspay/connector-service/commit/4006a76b00ada9e9038a57a49a97f09e3ac9c4ce))
+
+**Full Changelog:** [`2026.01.12.0...2026.01.12.1`](https://github.com/juspay/connector-service/compare/2026.01.12.0...2026.01.12.1)
+
+- - -
+
+## 2026.01.12.0
+
+### Features
+
+- **connector:** [Fiuu] Added RepeatPayment flow ([#414](https://github.com/juspay/connector-service/pull/414)) ([`b2f72d1`](https://github.com/juspay/connector-service/commit/b2f72d1e25d011dbb959a7d728e0ea57ed5a30ab))
+- **core:** MandateRevoke flow ([#214](https://github.com/juspay/connector-service/pull/214)) ([`b251e1a`](https://github.com/juspay/connector-service/commit/b251e1a490b472a896c26d85d416afd471ba5a9d))
+- **framework:** Added IncrementalAuthorization Flow support ([#410](https://github.com/juspay/connector-service/pull/410)) ([`be2fd45`](https://github.com/juspay/connector-service/commit/be2fd45b013ea6e04c2a58cfbbda028d6a71e19e))
+- Added Network-level error details in proto ([#417](https://github.com/juspay/connector-service/pull/417)) ([`772548a`](https://github.com/juspay/connector-service/commit/772548a7d4139906ec1abc4b45a4549f91d8c777))
+
+### Bug Fixes
+
+- Resolved RouterData diffs in Prod for Authorizedotnet ([#413](https://github.com/juspay/connector-service/pull/413)) ([`8d2fcb8`](https://github.com/juspay/connector-service/commit/8d2fcb8baea8b0ab59636d4d02c51a1d9fe061ff))
+
+**Full Changelog:** [`2026.01.09.0...2026.01.12.0`](https://github.com/juspay/connector-service/compare/2026.01.09.0...2026.01.12.0)
+
+- - -
+
 ## 2026.01.09.0
 
 ### Features